Integrated the Google API search
Now have a list of 10 results that come through with the desired characteristics from the requirements on query.

Also included a pagination capability.

Ended up doing this all on the front end, though additional separation could be accomplished through the use of library functions, did not do that for initial implementation.